Added RSS feed
Implemented an RSS feed
Added code to create new RSS items whenever an event or update us posted. or when the status of an event changes.
Moved the Bootstrap and AngularJS files into a folder in the website's root folder (rather than the parent folder as before)

committing all the files used in this app to ensure small changes aren't missed.